refactor(transcription): reorganize transcription service configurations
